addressing reviewer comments
